# Performance MetaDataEditor Gallery Drag-&-Drop via jQuery-UI