--- a/Kitodo/src/main/webapp/WEB-INF/templates/includes/metadataEditor/gallery.xhtml
+++ b/Kitodo/src/main/webapp/WEB-INF/templates/includes/metadataEditor/gallery.xhtml
@@ -24,36 +24,6 @@
         <!--@elvariable id="editImages" type="boolean"-->
         <ui:param name="editImages" value="#{SecurityAccessController.hasAuthorityToEditProcessImages()}"/>
 
-        <script>                
-            function makeDragAndDroppable() {
-                const startTime = Date.now();
-                $("#imagePreviewForm\\:structuredPages .draggableStructurePagePanel").draggable({
-                    scope: "assignedPagesDroppable",
-                    stack: ".ui-panel",
-                    revert: true,                   
-                });
-
-                $("#imagePreviewForm\\:structuredPages .page-drop-area").droppable({
-                    scope: "assignedPagesDroppable",
-                    activeClass: "media-stripe-index-active",        
-                    drop: function(event, ui) {
-                        let dragId = ui.draggable.attr('id');
-                        let dropId = $(event.target).attr('id');                       
-                        triggerOnPageDrop([{"name": "dragId", "value": dragId}, {"name": "dropId", "value": dropId}]);
-                    }         
-                });
-                console.log("makeDragAndDroppable in " + (Date.now() - startTime) + "ms");
-            }
-
-            var makeDragAndDroppableTimer = null;
-            var originalPrimeFacesAjaxUtilsUpdateElement = PrimeFaces.ajax.Utils.updateElement;
-            PrimeFaces.ajax.Utils.updateElement = function(id, content, xhr) {
-                originalPrimeFacesAjaxUtilsUpdateElement.apply(this, arguments);
-                clearTimeout(makeDragAndDroppableTimer);
-                makeDragAndDroppableTimer = setTimeout(makeDragAndDroppable, 250);
-            };
-        </script>
-
         <h:form id="imagePreviewForm" style="height: 100%;" styleClass="focusable">
 
             <p:remoteCommand 
@@ -175,40 +145,12 @@
                                         </p:outputPanel>
                                     </p:outputPanel>
                                 </p:panel>
-                                <!--<ui:fragment rendered="#{editImages}">
-                                    <p:draggable id="structuredPagesDraggable"
-                                                 scope="assignedPagesDroppable"
-                                                 for="imagePreviewForm:structuredPages:#{currentElement.rowIndex}:structureElementDataList:#{structuredThumbnail.rowIndex}:structuredPagePanel"
-                                                 revert="true"
-                                                 stack=".ui-panel"/>
-                                    <p:droppable for="imagePreviewForm:structuredPages:#{currentElement.rowIndex}:structureElementDataList:#{structuredThumbnail.rowIndex}:structuredPageDropArea"
-                                                 scope="assignedPagesDroppable"
-                                                 activeStyleClass="media-stripe-index-active">
-                                        <p:ajax listener="#{DataEditorForm.galleryPanel.onPageDrop}"
-                                                update="@(.pageList)"/>
-                                    </p:droppable>
-                                </ui:fragment>-->
                                 <!-- add one last drop area after the last page in a stripe -->
-                                <!--<ui:fragment rendered="#{editImages and stripe.medias.size() ne 0 and  stripe.medias.size() - 1 eq structuredThumbnail.rowIndex}">
+                                <ui:fragment rendered="#{editImages and stripe.medias.size() ne 0 and  stripe.medias.size() - 1 eq structuredThumbnail.rowIndex}">
                                     <p:panel id="structuredPageLastDropArea"
                                              styleClass="page-drop-area"/>
-                                    <p:droppable for="imagePreviewForm:structuredPages:#{currentElement.rowIndex}:structureElementDataList:#{structuredThumbnail.rowIndex}:structuredPageLastDropArea"
-                                                 scope="assignedPagesDroppable"
-                                                 activeStyleClass="media-stripe-index-active">
-                                        <p:ajax listener="#{DataEditorForm.galleryPanel.onPageDrop}"
-                                                update="@(.pageList)"/>
-                                    </p:droppable>
-                                </ui:fragment>-->
+                                </ui:fragment>
                             </p:dataList>
-                            <!--<ui:fragment rendered="#{editImages and stripe.medias.size() eq 0}">
-                                <p:droppable id="structuredPagesDroppable"
-                                             scope="assignedPagesDroppable"
-                                             for="imagePreviewForm:structuredPages:#{currentElement.rowIndex}:structureElementDataList"
-                                             activeStyleClass="media-stripe-active">
-                                    <p:ajax listener="#{DataEditorForm.galleryPanel.onPageDrop}"
-                                            update="@(.pageList)"/>
-                                </p:droppable>
-                            </ui:fragment>-->
                         </p:outputPanel>
                     </p:dataList>
 
@@ -232,7 +174,7 @@
                                     <p:panel id="unstructuredPageDropArea"
                                              rendered="#{editImages}"
                                              styleClass="page-drop-area"/>
-                                    <p:panel id="unstructuredMediaPanel">
+                                    <p:panel id="unstructuredMediaPanel" styleClass="draggableUnstructuredMediaPanel">
                                         <h:panelGroup id="updateSelectedUnstructuredMediaLink" styleClass="thumbnail-parent">
                                             <p:outputPanel styleClass="thumbnail #{DataEditorForm.consecutivePagesSelected() ? '' : 'discontinuous'} #{DataEditorForm.galleryPanel.isSelected(media, DataEditorForm.galleryPanel.stripes.get(0)) ? 'selected' : ''} #{DataEditorForm.galleryPanel.isLastSelection(media, DataEditorForm.galleryPanel.stripes.get(0)) ? 'last-selection' : ''}"/>
                                             <p:outputPanel styleClass="thumbnail-container"
@@ -254,40 +196,12 @@
                                             </p:outputPanel>
                                         </h:panelGroup>
                                     </p:panel>
-                                    <!--<ui:fragment rendered="#{editImages}">
-                                        <p:draggable id="unstructuredMediaDraggable"
-                                                     for="imagePreviewForm:unstructuredMediaList:#{currentMedia.rowIndex}:unstructuredMediaPanel"
-                                                     revert="true"
-                                                     scope="assignedPagesDroppable"
-                                                     stack=".ui-panel"/>
-                                        <p:droppable for="imagePreviewForm:unstructuredMediaList:#{currentMedia.rowIndex}:unstructuredPageDropArea"
-                                                     scope="assignedPagesDroppable"
-                                                     activeStyleClass="media-stripe-index-active">
-                                            <p:ajax listener="#{DataEditorForm.galleryPanel.onPageDrop}"
-                                                    update="@(.pageList)"/>
-                                        </p:droppable>
-                                    </ui:fragment>-->
                                     <!-- add one last drop area after the last page in a stripe -->
-                                    <!--<ui:fragment rendered="#{editImages and DataEditorForm.galleryPanel.stripes.get(0).medias.size() ne 0 and DataEditorForm.galleryPanel.stripes.get(0).medias.size() - 1 eq currentMedia.rowIndex}">
+                                    <ui:fragment rendered="#{editImages and DataEditorForm.galleryPanel.stripes.get(0).medias.size() ne 0 and DataEditorForm.galleryPanel.stripes.get(0).medias.size() - 1 eq currentMedia.rowIndex}">
                                         <p:panel id="unstructuredPageLastDropArea"
                                                  styleClass="page-drop-area"/>
-                                        <p:droppable for="imagePreviewForm:unstructuredMediaList:#{currentMedia.rowIndex}:unstructuredPageLastDropArea"
-                                                     scope="assignedPagesDroppable"
-                                                     activeStyleClass="media-stripe-index-active">
-                                            <p:ajax listener="#{DataEditorForm.galleryPanel.onPageDrop}"
-                                                    update="@(.pageList)"/>
-                                        </p:droppable>
-                                    </ui:fragment>-->
+                                    </ui:fragment>
                                 </p:dataList>
-                                <!--<ui:fragment rendered="#{editImages and DataEditorForm.galleryPanel.stripes.get(0).medias.size() eq 0}">
-                                    <p:droppable id="unstructuredPagesDroppable"
-                                                 scope="assignedPagesDroppable"
-                                                 for="imagePreviewForm:unstructuredMediaList"
-                                                 activeStyleClass="media-stripe-active">
-                                        <p:ajax listener="#{DataEditorForm.galleryPanel.onPageDrop}"
-                                                update="@(.pageList)"/>
-                                    </p:droppable>
-                                </ui:fragment>-->
                             </p:outputPanel>
                         </h:panelGroup>
                     </p:outputPanel>
@@ -480,5 +394,6 @@
         <h:outputStylesheet name="webjars/openlayers/5.2.0/ol.css"/>
         <h:outputScript library="webjars" name="openlayers/5.2.0/ol.js" />
         <h:outputScript name="js/ol_custom.js"/>
+        <h:outputScript name="js/metadataeditor_gallery_drag_drop.js" rendered="#{editImages}" />
     </p:panel>
 </ui:composition>