Related issues:
- #4195 

Conflicts with:
- #5011 

When navigating to the meta data editor, a lot of loading time is spent at initializing the drag-&-drop functionality of the gallery panel and structure tree. This can be tested by removing the authority to edit images of a process, which disables drag-&-drop in the meta data editor.

The problem can be traced back to the Primefaces components `<p:draggable />` and `<p:droppable />`. Unfortunately, they can only reference a single other component that is supposed to be made draggable or droppable via their `for="some-component-id"`-property. Because of this, and in order to make every thumbnail draggable, they have to be placed inside the `<p:dataList />` components and, thus, are repeated many times. Presumably, this causes the browser to initialize all event handlers for each drag-&-drop individually, which slows down the page drastically.

Luckily, Primefaces uses jQuery-UI in order to provide drag-&-drop functionality. This pull request removes `<p:draggable />` and `<p:droppable>` from the gallery panel and replaces it with corresponding Javascript code that initializes the drag-&-drop functionality via jQuery-UI directly. As a result, all thumbnails can be made draggable via one method call to jQuery-UI, which is much more efficient.

### Advantages

The loading time of a meta data editor with 1000 pages is reduced from 5.79 seconds to 3.33 seconds (desktop workstation) or 20.81 seconds to 9.94 seconds (laptop). 

Together with the following other pull requests
- #5114
- #5115 

the loading time is reduced from 5.79 seconds to 2.27 seconds (desktop workstation) or 20.81 seconds to 6.01 seconds (laptop).

### Drawbacks

**Compatibility**: Since components are made draggable outside of Primefaces, if those components are updated via Primefaces, they will loose all their custom event handlers. Therefore, a custom Javascript hook was implemented, which re-initializes drag-&-drop functionality after each update of any Primefaces components of the meta data editor. However, this does not slow down other parts of the meta data editor, because jQuery-UI is able to detect if a component is already draggable, and skip its re-initialization. Still, in the future, this hook or any other parts of the custom Javascript code may break in Primefaces v10 or v11.

**Maintainability**: Since this pull request removes Primefaces components and replaces them with custom Javascript code, the overall maintainability of the meta data editor decreases somewhat more. 

An alternative would be to reduce the number of draggable and droppable components via pagination as suggested in: 
- #5010