@markusweigelt 2022-04-29
@thomaslow The behavior of meta editor has changed a bit with this, since there are no more visual drop areas, right?

Is the problem also for larger number of pages fixed? Question from the following background:

For the discussion https://github.com/kitodo/kitodo-production/pull/5011 i created a PR which replaces primeface drag & drop too and implemented drag & drop with vanilla js instead of jquery (to reduce registered event handlers to one handler on document). Besides, the number of stucture elements has been reduced due to the conversion using CSS. PR was not completely finished because the behavior with more than 1500 pages or in specific test object with more than 1900 pages drag and drop was no longer possible.

Is drag & drop also possible for 1900 or more pages with this change? At that time I came to the conclusion that with higher number of stuctural elements and images drag & drop becomes slower and slower. Will this implementation make it better or do you also think that the editor only works up to a certain limit of pages and images?

In addition, there is the point that the client computers may not have as many resources available for the browser, as is the case with our development computers.

What do you mean?