@thomaslow 2022-04-29
@markusweigelt 
> @thomaslow The behavior of meta editor has changed a bit with this, since there are no more visual drop areas, right?

No, drop areas are preserved by this pull request. Each drop area is also initialized outside of Primefaces via a call to jquery-ui. From a user perspective, there should be no difference to the previous drag & drop behavior (except performance). I tested dragging of single pages, dragging to empty stripes and dragging of multiple selected pages.

> Is the problem also for larger number of pages fixed? Question from the following background:
>
> For the discussion #5011 i created a PR which replaces primeface drag & drop too and implemented drag & drop with vanilla js instead of jquery (to reduce registered event handlers to one handler on document). Besides, the number of stucture elements has been reduced due to the conversion using CSS. PR was not completely finished because the behavior with more than 1500 pages or in specific test object with more than 1900 pages drag and drop was no longer possible.

Unfortunately, I just discovered your PR. Maybe we should talk and see whether we can combine some changes?

> Is drag & drop also possible for 1900 or more pages with this change? At that time I came to the conclusion that with higher number of stuctural elements and images drag & drop becomes slower and slower. Will this implementation make it better or do you also think that the editor only works up to a certain limit of pages and images?

There is another issue with drag and drop, see #5116. With both changes, drag-&-drop is improved a lot, but not really nice to use. 

One problem that remains is that selecting a page in the gallery view triggers a synchronization between the gallery view and structure view. This happens right before dragging starts, and also when dropping a page, which causes another delay. 

> In addition, there is the point that the client computers may not have as many resources available for the browser, as is the case with our development computers.

I tested a process with 1000 pages and ~1500 pages and couldn't really find any resource problems on my development workstation. Maybe I should add some more pages.

Can you tell me what kind of limiting resources they have (RAM, CPU, etc.), and what browsers are used?