@markusweigelt 2022-05-02
@thomaslow 

> Unfortunately, I just discovered your PR. Maybe we should talk and see whether we can combine some changes?

This is not a problem and we do not need to do it. The PR was just a draft to show that the drag & drop changes alone can improve the performance but with increasing number of structural elements and images the processing is capped. e.g. to a maximum number of DOM elements that are possible and to the resources of the client. 

> There is another issue with drag and drop, see #5116. With both changes, drag-&-drop is improved a lot, but not really nice to use.

The loading behavior of Primeface is then an additional "annoying" factor that is true and if the fix remedies this is great.

> I tested a process with 1000 pages and ~1500 pages and couldn't really find any resource problems on my development workstation. Maybe I should add some more pages.
> 
> Can you tell me what kind of limiting resources they have (RAM, CPU, etc.), and what browsers are used?

I cannot make any more concrete statements on this. My only assumption here was that our developer computers are better than those in digitization and that problems are more likely to occur here.

In the concrete case example: there are >160 structural elements (chapters and subchapters) in the gallery on which > 1900 images/pages are distributed.  All the better, if this works great and there is still room for improvement. :)